HALS for irregular data
The HALS method in Processing now works on BP data that is not aligned with the GW data and thus, does not have to be regularly sampled either. To achieve this, all BP records that are within the same time period as the GW data are extracted and used for analysis at each location.
A short HALS notebook was created to demonstrate this.